Refactor out the function that acts as a GitHub App
To facilitate local debugging, let's factor out as much code into
modules as possible.

This commit is best viewed with `--color-moved`.
